extract backend to submodule
