+++ node/core/chain-selection/src/lib.rs
@@ -27,8 +27,11 @@ use polkadot_subsystem::{
 use parity_scale_codec::Error as CodecError;
 use futures::channel::oneshot;
 
-use std::collections::HashMap;
 use std::time::{UNIX_EPOCH, SystemTime};
+
+use crate::backend::{Backend, OverlayedBackend, BackendWriteOp};
+
+mod backend;
 
 const LOG_TARGET: &str = "parachain::chain-selection";
 
@@ -201,147 +204,6 @@ fn stagnant_timeout_from_now() -> Timestamp {
 	const STAGNANT_TIMEOUT: Timestamp = 120;
 
 	timestamp_now() + STAGNANT_TIMEOUT
-}
-
-enum BackendWriteOp {
-	WriteBlockEntry(BlockEntry),
-	WriteBlocksByNumber(BlockNumber, Vec<Hash>),
-	WriteViableLeaves(LeafEntrySet),
-	WriteStagnantAt(Timestamp, Vec<Hash>),
-	DeleteBlocksByNumber(BlockNumber),
-	DeleteBlockEntry(Hash),
-	DeleteStagnantAt(Timestamp),
-}
-
-// An abstraction over backend for the logic of this subsystem.
-// TODO [now]: extract to submodule
-trait Backend {
-	/// Load a block entry from the DB.
-	fn load_block_entry(&self, hash: &Hash) -> Result<Option<BlockEntry>, Error>;
-	/// Load the active-leaves set.
-	fn load_leaves(&self) -> Result<LeafEntrySet, Error>;
-	/// Load the stagnant list at the given timestamp.
-	fn load_stagnant_at(&self, timestamp: Timestamp) -> Result<Vec<Hash>, Error>;
-	/// Load all stagnant lists up to and including the given unix timestamp.
-	fn load_stagnant_at_up_to(&self, up_to: Timestamp)
-		-> Result<Vec<(Timestamp, Vec<Hash>)>, Error>;
-	/// Load the earliest kept block number.
-	fn load_first_block_number(&self) -> Result<Option<BlockNumber>, Error>;
-	/// Load blocks by number.
-	fn load_blocks_by_number(&self, number: BlockNumber) -> Result<Vec<Hash>, Error>;
-
-	/// Atomically write the list of operations, with later operations taking precedence over prior.
-	fn write(&mut self, ops: Vec<BackendWriteOp>) -> Result<(), Error>;
-}
-
-// An in-memory overlay over the backend.
-struct OverlayedBackend<'a, B: 'a> {
-	inner: &'a B,
-
-	// `None` means 'deleted', missing means query inner.
-	block_entries: HashMap<Hash, Option<BlockEntry>>,
-	// `None` means 'deleted', missing means query inner.
-	blocks_by_number: HashMap<BlockNumber, Option<Vec<Hash>>>,
-	// 'None' means 'deleted', missing means query inner.
-	stagnant_at: HashMap<Timestamp, Option<Vec<Hash>>>,
-	// 'None' means query inner.
-	leaves: Option<LeafEntrySet>,
-}
-
-impl<'a, B: 'a + Backend> OverlayedBackend<'a, B> {
-	fn new(backend: &'a B) -> Self {
-		OverlayedBackend {
-			inner: backend,
-			block_entries: HashMap::new(),
-			blocks_by_number: HashMap::new(),
-			stagnant_at: HashMap::new(),
-			leaves: None,
-		}
-	}
-
-	fn load_block_entry(&self, hash: &Hash) -> Result<Option<BlockEntry>, Error> {
-		if let Some(val) = self.block_entries.get(&hash) {
-			return Ok(val.clone())
-		}
-
-		self.inner.load_block_entry(hash)
-	}
-
-	fn load_blocks_by_number(&self, number: BlockNumber) -> Result<Vec<Hash>, Error> {
-		if let Some(val) = self.blocks_by_number.get(&number) {
-			return Ok(val.as_ref().map_or(Vec::new(), Clone::clone));
-		}
-
-		self.inner.load_blocks_by_number(number)
-	}
-
-	fn load_leaves(&self) -> Result<LeafEntrySet, Error> {
-		if let Some(ref set) = self.leaves {
-			return Ok(set.clone())
-		}
-
-		self.inner.load_leaves()
-	}
-
-	fn load_stagnant_at(&self, timestamp: Timestamp) -> Result<Vec<Hash>, Error> {
-		if let Some(val) = self.stagnant_at.get(&timestamp) {
-			return Ok(val.as_ref().map_or(Vec::new(), Clone::clone));
-		}
-
-		self.inner.load_stagnant_at(timestamp)
-	}
-
-	fn write_block_entry(&mut self, entry: BlockEntry) {
-		self.block_entries.insert(entry.block_hash, Some(entry));
-	}
-
-	fn delete_block_entry(&mut self, hash: &Hash) {
-		self.block_entries.remove(hash);
-	}
-
-	fn write_blocks_by_number(&mut self, number: BlockNumber, blocks: Vec<Hash>) {
-		self.blocks_by_number.insert(number, Some(blocks));
-	}
-
-	fn delete_blocks_by_number(&mut self, number: BlockNumber) {
-		self.blocks_by_number.insert(number, None);
-	}
-
-	fn write_leaves(&mut self, leaves: LeafEntrySet) {
-		self.leaves = Some(leaves);
-	}
-
-	fn write_stagnant_at(&mut self, timestamp: Timestamp, hashes: Vec<Hash>) {
-		self.stagnant_at.insert(timestamp, Some(hashes));
-	}
-
-	fn delete_stagnant_at(&mut self, timestamp: Timestamp) {
-		self.stagnant_at.insert(timestamp, None);
-	}
-
-	fn into_write_ops(self) -> impl Iterator<Item = BackendWriteOp> {
-		let block_entry_ops = self.block_entries.into_iter().map(|(h, v)| match v {
-			Some(v) => BackendWriteOp::WriteBlockEntry(v),
-			None => BackendWriteOp::DeleteBlockEntry(h),
-		});
-
-		let blocks_by_number_ops = self.blocks_by_number.into_iter().map(|(n, v)| match v {
-			Some(v) => BackendWriteOp::WriteBlocksByNumber(n, v),
-			None => BackendWriteOp::DeleteBlocksByNumber(n),
-		});
-
-		let leaf_ops = self.leaves.into_iter().map(BackendWriteOp::WriteViableLeaves);
-
-		let stagnant_at_ops = self.stagnant_at.into_iter().map(|(n, v)| match v {
-			Some(v) => BackendWriteOp::WriteStagnantAt(n, v),
-			None => BackendWriteOp::DeleteStagnantAt(n),
-		});
-
-		block_entry_ops
-			.chain(blocks_by_number_ops)
-			.chain(leaf_ops)
-			.chain(stagnant_at_ops)
-	}
 }
 
 async fn run<Context, B>(mut ctx: Context, mut backend: B)
